Fix byte index calculation bug.
This was a long-standing bug:

The index calculation was not checking properly for OOB cases where the
index is out of bounds of the prefetched slice. In such cases,
mistakenly, true was returned with a positive offset. This was causing
wrong data to be accessed if the access is to a non-prefetched index.

In the initial tests this was never bumped into as our prefetches thus
far have always been precise. But with imprecise prefetching this caused
some validation errors.

Arguably the newly added logic can/should be optimized. It would be nice
to compare the performance results pre- and post- this fix.